Add metrics fields to retro form
