+++ webapp/src/components/backstage/playbook_runs/playbook_run_backstage/playbook_run_backstage.tsx
@@ -436,6 +436,7 @@ const PlaybookRunBackstage = () => {
                         <Route path={`${match.url}/retrospective`}>
                             <Retrospective
                                 playbookRun={playbookRun}
+                                playbook={playbook}
                                 deleteTimelineEvent={deleteTimelineEvent}
                                 setRetrospective={setRetrospective}
                                 setPublishedAt={setPublishedAt}
