+++ webapp/src/components/backstage/playbook_runs/playbook_run_backstage/retrospective/report.tsx
@@ -1,82 +1,27 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {useState} from 'react';
-import styled from 'styled-components';
-import {useIntl, FormattedMessage} from 'react-intl';
-import debounce from 'debounce';
-import {DateTime} from 'luxon';
+import React from 'react';
 
-import {Timestamp} from 'src/webapp_globals';
+import styled from 'styled-components';
+import {useIntl} from 'react-intl';
+import debounce from 'debounce';
 
 import {PlaybookRun} from 'src/types/playbook_run';
-import {Title} from 'src/components/backstage/playbook_runs/shared';
-import {publishRetrospective, updateRetrospective} from 'src/client';
-import {PrimaryButton} from 'src/components/assets/buttons';
+import {updateRetrospective} from 'src/client';
 import ReportTextArea
     from 'src/components/backstage/playbook_runs/playbook_run_backstage/retrospective/report_text_area';
-import ConfirmModalLight from 'src/components/widgets/confirmation_modal_light';
 
 const editDebounceDelayMilliseconds = 2000;
 
 interface ReportProps {
     playbookRun: PlaybookRun;
+    isPublished: boolean;
     setRetrospective: (report: string) => void;
-    setPublishedAt: (publishedAt: number) => void;
-    setCanceled: (canceled: boolean) => void;
 }
 
-const PUB_TIME = {
-    useTime: false,
-    units: [
-        {within: ['second', -45], display: <FormattedMessage defaultMessage='just now'/>},
-        ['minute', -59],
-        ['hour', -48],
-        ['day', -30],
-        ['month', -12],
-        'year',
-    ],
-};
-
 const Report = (props: ReportProps) => {
-    const [showConfirmation, setShowConfirmation] = useState(false);
     const {formatMessage} = useIntl();
-
-    const confirmedPublish = () => {
-        publishRetrospective(props.playbookRun.id, props.playbookRun.retrospective);
-        props.setPublishedAt(DateTime.now().valueOf());
-        props.setCanceled(false);
-        setShowConfirmation(false);
-    };
-
-    const publishButtonText: React.ReactNode = formatMessage({defaultMessage: 'Publish'});
-    let publishComponent = (
-        <PrimaryButtonSmaller
-            onClick={() => setShowConfirmation(true)}
-        >
-            <TextContainer>{publishButtonText}</TextContainer>
-        </PrimaryButtonSmaller>
-    );
-
-    const isPublished = props.playbookRun.retrospective_published_at > 0 && !props.playbookRun.retrospective_was_canceled;
-    if (isPublished) {
-        const publishedAt = (
-            <Timestamp
-                value={props.playbookRun.retrospective_published_at}
-                {...PUB_TIME}
-            />
-        );
-        publishComponent = (
-            <>
-                <i className={'icon icon-check-all'}/>
-                <span>{''}</span>
-                {formatMessage({defaultMessage: 'Published {timestamp}'}, {timestamp: publishedAt})}
-                <DisabledPrimaryButtonSmaller>
-                    <TextContainer>{publishButtonText}</TextContainer>
-                </DisabledPrimaryButtonSmaller>
-            </>
-        );
-    }
 
     const persistEditEvent = (text: string) => {
         updateRetrospective(props.playbookRun.id, text);
@@ -88,24 +33,13 @@ const Report = (props: ReportProps) => {
         <ReportContainer>
             <Header>
                 <Title>{formatMessage({defaultMessage: 'Report'})}</Title>
-                <HeaderButtonsRight>
-                    {publishComponent}
-                </HeaderButtonsRight>
             </Header>
             <ReportTextArea
                 teamId={props.playbookRun.team_id}
                 text={props.playbookRun.retrospective}
-                isEditable={!isPublished}
+                isEditable={!props.isPublished}
                 onEdit={debouncedPersistEditEvent}
                 flushChanges={() => debouncedPersistEditEvent.flush()}
-            />
-            <ConfirmModalLight
-                show={showConfirmation}
-                title={formatMessage({defaultMessage: 'Are you sure you want to publish?'})}
-                message={formatMessage({defaultMessage: 'You will not be able to edit the retrospective report after publishing it. Do you want to publish the retrospective report?'})}
-                confirmButtonText={formatMessage({defaultMessage: 'Publish'})}
-                onConfirm={confirmedPublish}
-                onCancel={() => setShowConfirmation(false)}
             />
         </ReportContainer>
     );
@@ -116,50 +50,19 @@ const Header = styled.div`
     align-items: center;
 `;
 
-const HeaderButtonsRight = styled.div`
-    flex-grow: 1;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-
-    > * {
-        margin-left: 4px;
-    }
-`;
-
 const ReportContainer = styled.div`
     font-size: 12px;
     font-weight: normal;
     margin-bottom: 20px;
+    margin-top: 24px;
     height: 100%;
     display: flex;
     flex-direction: column;
 `;
 
-const PrimaryButtonSmaller = styled(PrimaryButton)`
-    height: 32px;
-`;
-
-const DisabledPrimaryButtonSmaller = styled(PrimaryButtonSmaller)`
-    background: rgba(var(--center-channel-color-rgb),0.08);
-    color: rgba(var(--center-channel-color-rgb),0.32);
-    margin-left: 16px;
-    cursor: default;
-
-    &:active:not([disabled])  {
-        background: rgba(var(--center-channel-color-rgb),0.08);
-    }
-
-    &:hover:enabled {
-        background: rgba(var(--center-channel-color-rgb),0.08);
-        &:before {
-            opacity: 0;
-        }
-    }
-`;
-
-const TextContainer = styled.span`
-    display: flex;
+const Title = styled.div`
+    font-weight: 600;
+    font-size: 14px;
 `;
 
 export default Report;
